+++ tests/conftest.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 import os
-from typing import Tuple, Callable
+from typing import Tuple, Callable, Optional
 
 from cellrank.ul.models import GAM, GAMR, SKLearnModel
 
@@ -143,12 +143,14 @@ def adata_cflare(adata_cflare=_create_cflare(backward=False)) -> AnnData:
 
 
 @pytest.fixture
-def gamr_model(adata_cflare: AnnData) -> GAMR:
-    m = GAMR(adata_cflare)
-    m.prepare(adata_cflare.var_names[0], "0").fit()
-    m.predict(level=0.95)
-
-    return m
+def gamr_model(adata_cflare: AnnData) -> Optional[GAMR]:
+    try:
+        m = GAMR(adata_cflare)
+        m.prepare(adata_cflare.var_names[0], "0").fit()
+        m.predict(level=0.95)
+        return m
+    except:
+        pytest.skip("Unable to create GAMR model.")
 
 
 @pytest.fixture